test(runner): organize doctor tests by behavior
