+++ crates/runner/src/cmd/doctor.rs
@@ -1284,1545 +1284,4 @@ fn format_uptime(started_at: &str) -> String {
 // ---------------------------------------------------------------------------
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use httpmock::prelude::*;
-
-    #[test]
-    fn format_uptime_minutes() {
-        let now = chrono::Utc::now();
-        let started = now - chrono::Duration::minutes(42);
-        let s = started.to_rfc3339();
-        assert_eq!(format_uptime(&s), "42m");
-    }
-
-    #[test]
-    fn format_uptime_hours_and_minutes() {
-        let now = chrono::Utc::now();
-        let started = now - chrono::Duration::hours(3) - chrono::Duration::minutes(15);
-        let s = started.to_rfc3339();
-        assert_eq!(format_uptime(&s), "3h 15m");
-    }
-
-    #[test]
-    fn format_uptime_days() {
-        let now = chrono::Utc::now();
-        let started = now - chrono::Duration::days(2) - chrono::Duration::hours(5);
-        let s = started.to_rfc3339();
-        assert_eq!(format_uptime(&s), "2d 5h 0m");
-    }
-
-    #[test]
-    fn format_uptime_invalid_timestamp() {
-        assert_eq!(format_uptime("not-a-date"), "unknown");
-    }
-
-    #[test]
-    fn parse_netns_list_line_valid() {
-        assert_eq!(
-            parse_netns_list_line("vm0-ns-00-0a (id: 42)"),
-            Some(("vm0-ns-00-0a", 0))
-        );
-        assert_eq!(
-            parse_netns_list_line("vm0-ns-3f-ff"),
-            Some(("vm0-ns-3f-ff", 63))
-        );
-    }
-
-    #[test]
-    fn parse_netns_list_line_ignores_malformed_names() {
-        assert_eq!(parse_netns_list_line(""), None);
-        assert_eq!(parse_netns_list_line("   "), None);
-        assert_eq!(parse_netns_list_line("not-a-ns"), None);
-        assert_eq!(parse_netns_list_line("vm0-ns-"), None);
-        assert_eq!(parse_netns_list_line("vm0-ns-00"), None);
-        assert_eq!(parse_netns_list_line("vm0-ns-00-extra"), None);
-        assert_eq!(parse_netns_list_line("vm0-ns-00-zz"), None);
-        assert_eq!(parse_netns_list_line("vm0-ns-00-0a-extra"), None);
-        assert_eq!(parse_netns_list_line("vm0-ns-0A-00"), None);
-        assert_eq!(parse_netns_list_line("vm0-ns-00-0A"), None);
-        assert_eq!(parse_netns_list_line("vm0-ns-40-00"), None);
-        assert_eq!(parse_netns_list_line("vm0-ns-ff-00"), None);
-    }
-
-    #[tokio::test]
-    async fn read_status_defaults_missing_collections_to_empty() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(
-            dir.path().join("status.json"),
-            r#"{"mode":"running","started_at":"2026-01-01T00:00:00.000Z"}"#,
-        )
-        .unwrap();
-
-        let status = read_status(dir.path()).await.unwrap();
-
-        assert!(status.active_runs.is_empty());
-        assert!(status.idle_vms.is_empty());
-    }
-
-    #[tokio::test]
-    async fn read_status_missing_active_run_identifier_returns_none() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(
-            dir.path().join("status.json"),
-            r#"{
-                "mode":"running",
-                "started_at":"2026-01-01T00:00:00.000Z",
-                "active_runs":[{"run_id":"R1"}]
-            }"#,
-        )
-        .unwrap();
-
-        assert!(read_status(dir.path()).await.is_none());
-    }
-
-    #[tokio::test]
-    async fn read_status_missing_idle_vm_identifier_returns_none() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(
-            dir.path().join("status.json"),
-            r#"{
-                "mode":"running",
-                "started_at":"2026-01-01T00:00:00.000Z",
-                "idle_vms":[{"session_id":"session-a"}]
-            }"#,
-        )
-        .unwrap();
-
-        assert!(read_status(dir.path()).await.is_none());
-    }
-
-    #[test]
-    fn active_run_unknown_phase_defaults_running() {
-        let active = ActiveRun {
-            run_id: "R1".into(),
-            sandbox_id: "S1".into(),
-            phase: Some("future-phase".into()),
-            phase_started_at: None,
-        };
-
-        assert_eq!(active.phase(), ActiveRunPhase::Running);
-    }
-
-    fn legacy_active_run(run_id: &str, sandbox_id: &str) -> ActiveRun {
-        ActiveRun {
-            run_id: run_id.into(),
-            sandbox_id: sandbox_id.into(),
-            phase: None,
-            phase_started_at: None,
-        }
-    }
-
-    fn phased_active_run(
-        run_id: &str,
-        sandbox_id: &str,
-        phase: &str,
-        phase_started_at: Option<String>,
-    ) -> ActiveRun {
-        ActiveRun {
-            run_id: run_id.into(),
-            sandbox_id: sandbox_id.into(),
-            phase: Some(phase.into()),
-            phase_started_at,
-        }
-    }
-
-    fn phase_started_at_ago(age: Duration) -> String {
-        let age = chrono::Duration::from_std(age).unwrap();
-        (Utc::now() - age)
-            .format("%Y-%m-%dT%H:%M:%S%.3fZ")
-            .to_string()
-    }
-
-    fn status_info(active: Vec<(&str, &str)>, idle_sandboxes: Vec<&str>) -> StatusInfo {
-        status_info_with_active_runs(
-            active
-                .into_iter()
-                .map(|(run_id, sandbox_id)| legacy_active_run(run_id, sandbox_id))
-                .collect(),
-            idle_sandboxes,
-        )
-    }
-
-    fn status_info_with_active_runs(
-        active_runs: Vec<ActiveRun>,
-        idle_sandboxes: Vec<&str>,
-    ) -> StatusInfo {
-        StatusInfo {
-            mode: "running".into(),
-            started_at: "2026-01-01T00:00:00.000Z".into(),
-            active_runs,
-            // Tests only need sandbox_id lookup for idle VMs; synthesize a
-            // placeholder session_id.
-            idle_vms: idle_sandboxes
-                .into_iter()
-                .enumerate()
-                .map(|(i, sbid)| IdleVm {
-                    session_id: format!("sess-{i}"),
-                    sandbox_id: sbid.into(),
-                })
-                .collect(),
-            proxy_port: None,
-            dns_port: None,
-        }
-    }
-
-    fn fc_info(pid: u32, sandbox_id: &str, base_dir: &str) -> process::FirecrackerProcessInfo {
-        process::FirecrackerProcessInfo {
-            pid,
-            ppid: None,
-            sandbox_id: sandbox_id.into(),
-            base_dir: Some(PathBuf::from(base_dir)),
-            identity: None,
-        }
-    }
-
-    #[test]
-    fn correlate_jobs_matching() {
-        // sandbox_id == run_id (first-job case: doctor still joins correctly).
-        let status = status_info(vec![("abc", "abc"), ("def", "def")], vec![]);
-        let fc = vec![
-            fc_info(100, "abc", "/data/r1"),
-            fc_info(101, "def", "/data/r1"),
-        ];
-        let (jobs, warnings) = correlate_jobs(&status, Path::new("/data/r1"), &fc);
-        assert_eq!(jobs.len(), 2);
-        assert!(warnings.is_empty());
-        assert!(matches!(
-            jobs.first().unwrap(),
-            JobStatus::Running { pid: 100, .. }
-        ));
-    }
-
-    #[test]
-    fn correlate_active_reused_no_warning() {
-        // Sandbox was reused: FC's sandbox_id differs from the active run_id
-        // but matches via the status mapping — should emit zero warnings.
-        let status = status_info(vec![("run-new", "sandbox-orig")], vec![]);
-        let fc = vec![fc_info(200, "sandbox-orig", "/data/r1")];
-        let (jobs, warnings) = correlate_jobs(&status, Path::new("/data/r1"), &fc);
-        assert_eq!(jobs.len(), 1);
-        assert!(warnings.is_empty(), "reused sandbox must not warn");
-        let JobStatus::Running { run_id, pid } = jobs.first().unwrap() else {
-            panic!("expected Running");
-        };
-        assert_eq!(
-            run_id, "run-new",
-            "Running should carry the run_id, not sandbox_id"
-        );
-        assert_eq!(*pid, 200);
-    }
-
-    #[test]
-    fn correlate_idle_vm_no_warning() {
-        // Parked idle VM: no active run, sandbox_id listed in idle_vms.
-        // This is the exact prod-3 v0.79.12 scenario that used to
-        // false-positive with 2 `FirecrackerNotInStatus` warnings.
-        let status = status_info(vec![], vec!["sandbox-idle"]);
-        let fc = vec![fc_info(300, "sandbox-idle", "/data/r1")];
-        let (jobs, warnings) = correlate_jobs(&status, Path::new("/data/r1"), &fc);
-        assert!(jobs.is_empty(), "idle FC should not surface as a job");
-        assert!(warnings.is_empty(), "idle FC must not warn");
-    }
-
-    #[test]
-    fn correlate_active_no_fc_warns() {
-        // Active run with no matching FC process.
-        let status = status_info(vec![("run-x", "sandbox-x")], vec![]);
-        let fc: Vec<process::FirecrackerProcessInfo> = vec![];
-        let (jobs, warnings) = correlate_jobs(&status, Path::new("/data/r1"), &fc);
-        assert_eq!(jobs.len(), 1);
-        assert!(matches!(jobs.first().unwrap(), JobStatus::NoProcess { .. }));
-        assert_eq!(warnings.len(), 1);
-        let msg = warnings[0].to_string();
-        assert!(msg.contains("no firecracker"), "{msg}");
-        assert!(msg.contains("run-x"), "{msg}");
-        assert!(msg.contains("sandbox-x"), "{msg}");
-    }
-
-    #[test]
-    fn correlate_preparing_active_without_fc_within_grace_does_not_warn() {
-        let status = status_info_with_active_runs(
-            vec![phased_active_run(
-                "run-prep",
-                "sandbox-prep",
-                "preparing",
-                Some(phase_started_at_ago(Duration::from_secs(5))),
-            )],
-            vec![],
-        );
-        let fc: Vec<process::FirecrackerProcessInfo> = vec![];
-
-        let (jobs, warnings) = correlate_jobs(&status, Path::new("/data/r1"), &fc);
-
-        assert_eq!(jobs.len(), 1);
-        assert!(warnings.is_empty());
-        assert!(matches!(jobs.first().unwrap(), JobStatus::Preparing { .. }));
-    }
-
-    #[test]
-    fn correlate_preparing_active_without_fc_beyond_grace_warns() {
-        let status = status_info_with_active_runs(
-            vec![phased_active_run(
-                "run-stale",
-                "sandbox-stale",
-                "preparing",
-                Some(phase_started_at_ago(
-                    PREPARING_NO_PROCESS_GRACE + Duration::from_secs(1),
-                )),
-            )],
-            vec![],
-        );
-        let fc: Vec<process::FirecrackerProcessInfo> = vec![];
-
-        let (jobs, warnings) = correlate_jobs(&status, Path::new("/data/r1"), &fc);
-
-        assert_eq!(jobs.len(), 1);
-        assert!(matches!(
-            jobs.first().unwrap(),
-            JobStatus::Preparing { pid: None, .. }
-        ));
-        assert_eq!(warnings.len(), 1);
-        let msg = warnings[0].to_string();
-        assert!(msg.contains("stuck preparing"), "{msg}");
-        assert!(msg.contains("run-stale"), "{msg}");
-        assert!(msg.contains("sandbox-stale"), "{msg}");
-    }
-
-    #[test]
-    fn correlate_preparing_active_with_fc_within_grace_stays_preparing() {
-        let status = status_info_with_active_runs(
-            vec![phased_active_run(
-                "run-prep",
-                "sandbox-prep",
-                "preparing",
-                Some(phase_started_at_ago(Duration::from_secs(5))),
-            )],
-            vec![],
-        );
-        let fc = vec![fc_info(123, "sandbox-prep", "/data/r1")];
-
-        let (jobs, warnings) = correlate_jobs(&status, Path::new("/data/r1"), &fc);
-
-        assert_eq!(jobs.len(), 1);
-        assert!(warnings.is_empty());
-        let JobStatus::Preparing { run_id, pid } = jobs.first().unwrap() else {
-            panic!("expected preparing job");
-        };
-        assert_eq!(run_id, "run-prep");
-        assert_eq!(*pid, Some(123));
-    }
-
-    #[test]
-    fn correlate_preparing_active_with_fc_beyond_grace_warns() {
-        let status = status_info_with_active_runs(
-            vec![phased_active_run(
-                "run-stale-with-fc",
-                "sandbox-stale-with-fc",
-                "preparing",
-                Some(phase_started_at_ago(
-                    PREPARING_NO_PROCESS_GRACE + Duration::from_secs(1),
-                )),
-            )],
-            vec![],
-        );
-        let fc = vec![fc_info(456, "sandbox-stale-with-fc", "/data/r1")];
-
-        let (jobs, warnings) = correlate_jobs(&status, Path::new("/data/r1"), &fc);
-
-        assert_eq!(jobs.len(), 1);
-        let JobStatus::Preparing { run_id, pid } = jobs.first().unwrap() else {
-            panic!("expected preparing job");
-        };
-        assert_eq!(run_id, "run-stale-with-fc");
-        assert_eq!(*pid, Some(456));
-        assert_eq!(warnings.len(), 1);
-        let msg = warnings[0].to_string();
-        assert!(msg.contains("stuck preparing"), "{msg}");
-        assert!(msg.contains("run-stale-with-fc"), "{msg}");
-        assert!(msg.contains("sandbox-stale-with-fc"), "{msg}");
-    }
-
-    #[test]
-    fn correlate_preparing_active_without_timestamp_warns() {
-        let status = status_info_with_active_runs(
-            vec![phased_active_run(
-                "run-missing-ts",
-                "sandbox-missing-ts",
-                "preparing",
-                None,
-            )],
-            vec![],
-        );
-        let fc: Vec<process::FirecrackerProcessInfo> = vec![];
-
-        let (jobs, warnings) = correlate_jobs(&status, Path::new("/data/r1"), &fc);
-
-        assert_eq!(jobs.len(), 1);
-        assert!(matches!(
-            jobs.first().unwrap(),
-            JobStatus::Preparing { pid: None, .. }
-        ));
-        assert_eq!(warnings.len(), 1);
-    }
-
-    #[test]
-    fn correlate_orphan_fc_warns() {
-        // FC process with a sandbox_id outside both active and idle sets.
-        let status = status_info(vec![("run-a", "sandbox-a")], vec!["sandbox-idle"]);
-        let fc = vec![
-            fc_info(400, "sandbox-a", "/data/r1"),      // active, OK
-            fc_info(401, "sandbox-idle", "/data/r1"),   // idle, OK
-            fc_info(402, "sandbox-orphan", "/data/r1"), // orphan, warn
-        ];
-        let (jobs, warnings) = correlate_jobs(&status, Path::new("/data/r1"), &fc);
-        assert_eq!(jobs.len(), 2, "one active job + one orphan surface");
-        assert_eq!(warnings.len(), 1);
-        let msg = warnings[0].to_string();
-        assert!(msg.contains("sandbox-orphan"), "{msg}");
-        assert!(msg.contains("not in status.json"), "{msg}");
-
-        // The orphan row must carry the sandbox_id in `NotInStatus`, not
-        // misfiled into a `run_id` variant. Type-checking this here locks
-        // in the distinction so future refactors can't quietly regress.
-        let orphan_row = jobs
-            .iter()
-            .find(|j| matches!(j, JobStatus::NotInStatus { .. }))
-            .expect("orphan row missing");
-        let JobStatus::NotInStatus { sandbox_id } = orphan_row else {
-            panic!("orphan row must be NotInStatus");
-        };
-        assert_eq!(sandbox_id, "sandbox-orphan");
-    }
-
-    #[test]
-    fn correlate_jobs_ignores_other_runners() {
-        let status = status_info(vec![("abc", "sbox-abc")], vec![]);
-        // This firecracker belongs to a different runner (different base_dir)
-        let fc = vec![process::FirecrackerProcessInfo {
-            pid: 300,
-            ppid: None,
-            sandbox_id: "sbox-abc".into(),
-            base_dir: Some(PathBuf::from("/data/r2")),
-            identity: None,
-        }];
-        let (jobs, warnings) = correlate_jobs(&status, Path::new("/data/r1"), &fc);
-        assert_eq!(jobs.len(), 1);
-        assert_eq!(warnings.len(), 1);
-        assert!(warnings[0].to_string().contains("no firecracker process"));
-    }
-
-    #[test]
-    fn correlate_jobs_empty_status_empty_fcs() {
-        let status = status_info(vec![], vec![]);
-        let fc: Vec<process::FirecrackerProcessInfo> = vec![];
-        let (jobs, warnings) = correlate_jobs(&status, Path::new("/data/r1"), &fc);
-        assert!(jobs.is_empty());
-        assert!(warnings.is_empty());
-    }
-
-    // -- persists() recheck tests --------------------------------------------
-    //
-    // These exercise the resolution logic used in the recheck loop. Each
-    // writes a real status.json (since persists() calls read_status
-    // internally) and asserts whether the warning clears.
-
-    fn write_status_json(path: &std::path::Path, contents: &str) {
-        std::fs::write(path, contents).unwrap();
-    }
-
-    fn empty_fresh() -> process::DiscoveredProcesses {
-        process::DiscoveredProcesses {
-            firecrackers: vec![],
-            mitmdumps: vec![],
-            dnsmasqs: vec![],
-        }
-    }
-
-    fn fresh_with_firecracker(
-        pid: u32,
-        sandbox_id: &str,
-        base_dir: &Path,
-    ) -> process::DiscoveredProcesses {
-        process::DiscoveredProcesses {
-            firecrackers: vec![process::FirecrackerProcessInfo {
-                pid,
-                ppid: None,
-                sandbox_id: sandbox_id.into(),
-                base_dir: Some(base_dir.to_path_buf()),
-                identity: None,
-            }],
-            mitmdumps: vec![],
-            dnsmasqs: vec![],
-        }
-    }
-
-    #[tokio::test]
-    async fn no_firecracker_for_run_resolves_when_run_removed_even_if_sandbox_reused() {
-        // Regression: a NoFirecrackerForRun warning for run R1 should clear
-        // once R1 is no longer in active_runs, even if R1's sandbox_id got
-        // reused by another active run R2. Keying on sandbox_id instead of
-        // run_id would wrongly keep R1's warning.
-        let dir = tempfile::tempdir().unwrap();
-        let base_dir = dir.path().to_path_buf();
-        write_status_json(
-            &base_dir.join("status.json"),
-            r#"{
-                "mode": "running",
-                "max_concurrent": 4,
-                "active_runs": [
-                    {"run_id": "R2", "sandbox_id": "S1"}
-                ],
-                "started_at": "2026-01-01T00:00:00.000Z",
-                "updated_at": "2026-01-01T00:00:00.000Z"
-            }"#,
-        );
-
-        let warning = Warning::NoFirecrackerForRun {
-            run_id: "R1".into(),
-            sandbox_id: "S1".into(),
-            base_dir: base_dir.clone(),
-        };
-        assert!(
-            !warning.persists(&empty_fresh(), &[]).await,
-            "warning about R1 must clear after R1 leaves active_runs even though S1 is reused"
-        );
-    }
-
-    #[tokio::test]
-    async fn no_firecracker_for_run_persists_while_run_still_active() {
-        let dir = tempfile::tempdir().unwrap();
-        let base_dir = dir.path().to_path_buf();
-        write_status_json(
-            &base_dir.join("status.json"),
-            r#"{
-                "mode": "running",
-                "max_concurrent": 4,
-                "active_runs": [
-                    {"run_id": "R1", "sandbox_id": "S1"}
-                ],
-                "started_at": "2026-01-01T00:00:00.000Z",
-                "updated_at": "2026-01-01T00:00:00.000Z"
-            }"#,
-        );
-
-        let warning = Warning::NoFirecrackerForRun {
-            run_id: "R1".into(),
-            sandbox_id: "S1".into(),
-            base_dir: base_dir.clone(),
-        };
-        // R1 is still active and there's no FC in fresh. Warning persists.
-        assert!(warning.persists(&empty_fresh(), &[]).await);
-    }
-
-    #[tokio::test]
-    async fn no_firecracker_for_run_clears_for_preparing_run_within_grace() {
-        let dir = tempfile::tempdir().unwrap();
-        let base_dir = dir.path().to_path_buf();
-        let phase_started_at = phase_started_at_ago(Duration::from_secs(10));
-        write_status_json(
-            &base_dir.join("status.json"),
-            &format!(
-                r#"{{
-                    "mode": "running",
-                    "max_concurrent": 4,
-                    "active_runs": [
-                        {{
-                            "run_id": "R1",
-                            "sandbox_id": "S1",
-                            "phase": "preparing",
-                            "phase_started_at": "{phase_started_at}"
-                        }}
-                    ],
-                    "started_at": "2026-01-01T00:00:00.000Z",
-                    "updated_at": "2026-01-01T00:00:00.000Z"
-                }}"#
-            ),
-        );
-
-        let warning = Warning::NoFirecrackerForRun {
-            run_id: "R1".into(),
-            sandbox_id: "S1".into(),
-            base_dir: base_dir.clone(),
-        };
-        assert!(!warning.persists(&empty_fresh(), &[]).await);
-    }
-
-    #[tokio::test]
-    async fn no_firecracker_for_run_clears_for_stale_preparing_run() {
-        let dir = tempfile::tempdir().unwrap();
-        let base_dir = dir.path().to_path_buf();
-        let phase_started_at =
-            phase_started_at_ago(PREPARING_NO_PROCESS_GRACE + Duration::from_secs(1));
-        write_status_json(
-            &base_dir.join("status.json"),
-            &format!(
-                r#"{{
-                    "mode": "running",
-                    "max_concurrent": 4,
-                    "active_runs": [
-                        {{
-                            "run_id": "R1",
-                            "sandbox_id": "S1",
-                            "phase": "preparing",
-                            "phase_started_at": "{phase_started_at}"
-                        }}
-                    ],
-                    "started_at": "2026-01-01T00:00:00.000Z",
-                    "updated_at": "2026-01-01T00:00:00.000Z"
-                }}"#
-            ),
-        );
-
-        let warning = Warning::NoFirecrackerForRun {
-            run_id: "R1".into(),
-            sandbox_id: "S1".into(),
-            base_dir: base_dir.clone(),
-        };
-        assert!(!warning.persists(&empty_fresh(), &[]).await);
-    }
-
-    #[tokio::test]
-    async fn stale_preparing_run_persists_while_run_remains_stale_preparing() {
-        let dir = tempfile::tempdir().unwrap();
-        let base_dir = dir.path().to_path_buf();
-        let phase_started_at =
-            phase_started_at_ago(PREPARING_NO_PROCESS_GRACE + Duration::from_secs(1));
-        write_status_json(
-            &base_dir.join("status.json"),
-            &format!(
-                r#"{{
-                    "mode": "running",
-                    "max_concurrent": 4,
-                    "active_runs": [
-                        {{
-                            "run_id": "R1",
-                            "sandbox_id": "S1",
-                            "phase": "preparing",
-                            "phase_started_at": "{phase_started_at}"
-                        }}
-                    ],
-                    "started_at": "2026-01-01T00:00:00.000Z",
-                    "updated_at": "2026-01-01T00:00:00.000Z"
-                }}"#
-            ),
-        );
-
-        let warning = Warning::StalePreparingRun {
-            run_id: "R1".into(),
-            sandbox_id: "S1".into(),
-            base_dir: base_dir.clone(),
-        };
-        assert!(warning.persists(&empty_fresh(), &[]).await);
-    }
-
-    #[tokio::test]
-    async fn stale_preparing_run_clears_after_phase_changes_to_running_with_firecracker() {
-        let dir = tempfile::tempdir().unwrap();
-        let base_dir = dir.path().to_path_buf();
-        write_status_json(
-            &base_dir.join("status.json"),
-            r#"{
-                "mode": "running",
-                "max_concurrent": 4,
-                "active_runs": [
-                    {
-                        "run_id": "R1",
-                        "sandbox_id": "S1",
-                        "phase": "running",
-                        "phase_started_at": "2026-01-01T00:00:00.000Z"
-                    }
-                ],
-                "started_at": "2026-01-01T00:00:00.000Z",
-                "updated_at": "2026-01-01T00:00:00.000Z"
-            }"#,
-        );
-
-        let warning = Warning::StalePreparingRun {
-            run_id: "R1".into(),
-            sandbox_id: "S1".into(),
-            base_dir: base_dir.clone(),
-        };
-        assert!(
-            !warning
-                .persists(&fresh_with_firecracker(123, "S1", &base_dir), &[])
-                .await
-        );
-    }
-
-    #[tokio::test]
-    async fn stale_preparing_run_persists_after_phase_changes_to_running_without_firecracker() {
-        let dir = tempfile::tempdir().unwrap();
-        let base_dir = dir.path().to_path_buf();
-        write_status_json(
-            &base_dir.join("status.json"),
-            r#"{
-                "mode": "running",
-                "max_concurrent": 4,
-                "active_runs": [
-                    {
-                        "run_id": "R1",
-                        "sandbox_id": "S1",
-                        "phase": "running",
-                        "phase_started_at": "2026-01-01T00:00:00.000Z"
-                    }
-                ],
-                "started_at": "2026-01-01T00:00:00.000Z",
-                "updated_at": "2026-01-01T00:00:00.000Z"
-            }"#,
-        );
-
-        let warning = Warning::StalePreparingRun {
-            run_id: "R1".into(),
-            sandbox_id: "S1".into(),
-            base_dir: base_dir.clone(),
-        };
-        assert!(warning.persists(&empty_fresh(), &[]).await);
-    }
-
-    #[tokio::test]
-    async fn no_firecracker_for_run_clears_when_same_run_points_to_new_sandbox() {
-        let dir = tempfile::tempdir().unwrap();
-        let base_dir = dir.path().to_path_buf();
-        write_status_json(
-            &base_dir.join("status.json"),
-            r#"{
-                "mode": "running",
-                "max_concurrent": 4,
-                "active_runs": [
-                    {"run_id": "R1", "sandbox_id": "S2"}
-                ],
-                "started_at": "2026-01-01T00:00:00.000Z",
-                "updated_at": "2026-01-01T00:00:00.000Z"
-            }"#,
-        );
-
-        let warning = Warning::NoFirecrackerForRun {
-            run_id: "R1".into(),
-            sandbox_id: "S1".into(),
-            base_dir: base_dir.clone(),
-        };
-        assert!(!warning.persists(&empty_fresh(), &[]).await);
-    }
-
-    #[tokio::test]
-    async fn firecracker_not_in_status_clears_when_sandbox_becomes_idle() {
-        // A FirecrackerNotInStatus warning must clear once its sandbox_id
-        // shows up in idle_vms (the VM was parked between scans).
-        let dir = tempfile::tempdir().unwrap();
-        let base_dir = dir.path().to_path_buf();
-        write_status_json(
-            &base_dir.join("status.json"),
-            r#"{
-                "mode": "running",
-                "max_concurrent": 4,
-                "active_runs": [],
-                "idle_vms": [
-                    {"session_id": "sess-1", "sandbox_id": "S1"}
-                ],
-                "started_at": "2026-01-01T00:00:00.000Z",
-                "updated_at": "2026-01-01T00:00:00.000Z"
-            }"#,
-        );
-
-        // Use our own PID (which certainly exists) so the pid_exists check
-        // doesn't short-circuit.
-        let live_pid = std::process::id();
-        let warning = Warning::FirecrackerNotInStatus {
-            pid: live_pid,
-            sandbox_id: "S1".into(),
-            base_dir: base_dir.clone(),
-        };
-        assert!(
-            !warning.persists(&empty_fresh(), &[]).await,
-            "warning must clear once the sandbox is tracked as idle"
-        );
-    }
-
-    #[tokio::test]
-    async fn firecracker_not_in_status_clears_when_sandbox_becomes_active() {
-        let dir = tempfile::tempdir().unwrap();
-        let base_dir = dir.path().to_path_buf();
-        write_status_json(
-            &base_dir.join("status.json"),
-            r#"{
-                "mode": "running",
-                "max_concurrent": 4,
-                "active_runs": [
-                    {"run_id": "R1", "sandbox_id": "S1"}
-                ],
-                "started_at": "2026-01-01T00:00:00.000Z",
-                "updated_at": "2026-01-01T00:00:00.000Z"
-            }"#,
-        );
-
-        let live_pid = std::process::id();
-        let warning = Warning::FirecrackerNotInStatus {
-            pid: live_pid,
-            sandbox_id: "S1".into(),
-            base_dir: base_dir.clone(),
-        };
-        assert!(!warning.persists(&empty_fresh(), &[]).await);
-    }
-
-    #[tokio::test]
-    async fn firecracker_not_in_status_persists_when_still_orphan() {
-        let dir = tempfile::tempdir().unwrap();
-        let base_dir = dir.path().to_path_buf();
-        write_status_json(
-            &base_dir.join("status.json"),
-            r#"{
-                "mode": "running",
-                "max_concurrent": 4,
-                "active_runs": [],
-                "started_at": "2026-01-01T00:00:00.000Z",
-                "updated_at": "2026-01-01T00:00:00.000Z"
-            }"#,
-        );
-
-        let live_pid = std::process::id();
-        let warning = Warning::FirecrackerNotInStatus {
-            pid: live_pid,
-            sandbox_id: "S-ghost".into(),
-            base_dir: base_dir.clone(),
-        };
-        assert!(warning.persists(&empty_fresh(), &[]).await);
-    }
-
-    #[tokio::test]
-    async fn firecracker_not_in_status_clears_when_pid_gone() {
-        let dir = tempfile::tempdir().unwrap();
-        let base_dir = dir.path().to_path_buf();
-        // status.json doesn't matter — pid check short-circuits.
-        let warning = Warning::FirecrackerNotInStatus {
-            pid: u32::MAX, // never a valid pid
-            sandbox_id: "S-anything".into(),
-            base_dir,
-        };
-        assert!(!warning.persists(&empty_fresh(), &[]).await);
-    }
-
-    #[tokio::test]
-    async fn orphan_mitmdump_clears_when_parent_runner_pid_appears() {
-        struct ChildGuard(std::process::Child);
-
-        impl Drop for ChildGuard {
-            fn drop(&mut self) {
-                let _ = self.0.kill();
-                let _ = self.0.wait();
-            }
-        }
-
-        let child = ChildGuard(
-            std::process::Command::new("sleep")
-                .arg("30")
-                .spawn()
-                .unwrap(),
-        );
-        let warning = Warning::OrphanMitmdump {
-            pid: child.0.id(),
-            port: 32821,
-            ppid: Some(std::process::id()),
-        };
-
-        assert!(warning.persists(&empty_fresh(), &[]).await);
-        assert!(
-            !warning
-                .persists(&empty_fresh(), &[std::process::id()])
-                .await
-        );
-    }
-
-    #[test]
-    fn find_stopped_services_detects_missing() {
-        let installed = vec![
-            InstalledService {
-                unit_name: "vm0-runner-active".into(),
-                config_path: Some(PathBuf::from("/data/active.yaml")),
-            },
-            InstalledService {
-                unit_name: "vm0-runner-stopped".into(),
-                config_path: Some(PathBuf::from("/data/stopped.yaml")),
-            },
-        ];
-        let reports = vec![RunnerReport {
-            live_runner: live_runner_instance(
-                1,
-                PathBuf::from("/data/active.yaml"),
-                PathBuf::from("/data/active"),
-            ),
-            service_type: ServiceType::Installed("vm0-runner-active".into()),
-            status: None,
-            api_ok: None,
-            proxy_pid: None,
-            dns_pid: None,
-            jobs: vec![],
-            warnings: vec![],
-        }];
-        let stopped = find_stopped_services(&installed, &reports);
-        assert_eq!(stopped.len(), 1);
-        assert_eq!(stopped[0].unit_name, "vm0-runner-stopped");
-        assert_eq!(stopped[0].config_info, "/data/stopped.yaml");
-    }
-
-    #[test]
-    fn warning_display() {
-        let w = Warning::NoFirecrackerForRun {
-            run_id: "abc-123".into(),
-            sandbox_id: "sbox-abc".into(),
-            base_dir: PathBuf::from("/data/r1"),
-        };
-        assert_eq!(
-            w.to_string(),
-            "no firecracker process for run abc-123 (sandbox sbox-abc)"
-        );
-
-        let w = Warning::StalePreparingRun {
-            run_id: "prep-123".into(),
-            sandbox_id: "sbox-prep".into(),
-            base_dir: PathBuf::from("/data/r1"),
-        };
-        assert_eq!(
-            w.to_string(),
-            "run prep-123 stuck preparing (sandbox sbox-prep)"
-        );
-
-        let w = Warning::FirecrackerNotInStatus {
-            pid: 17,
-            sandbox_id: "sbox-gone".into(),
-            base_dir: PathBuf::from("/data/r1"),
-        };
-        assert_eq!(
-            w.to_string(),
-            "firecracker PID 17 (sandbox sbox-gone) not in status.json"
-        );
-
-        let w = Warning::ApiUnreachable {
-            server_url: "https://example.com".into(),
-            server_token: "tok".into(),
-        };
-        assert_eq!(w.to_string(), "API unreachable");
-
-        let w = Warning::OrphanFirecracker {
-            pid: 42,
-            sandbox_id: "xyz".into(),
-            ppid: Some(10),
-        };
-        assert_eq!(
-            w.to_string(),
-            "orphan firecracker PID 42 (sandbox xyz, ppid=10)"
-        );
-
-        let w = Warning::OrphanFirecracker {
-            pid: 42,
-            sandbox_id: "xyz".into(),
-            ppid: None,
-        };
-        assert_eq!(
-            w.to_string(),
-            "orphan firecracker PID 42 (sandbox xyz, ppid=?)"
-        );
-
-        let w = Warning::StaleMitmproxy {
-            pid: 555,
-            port: 32821,
-        };
-        assert_eq!(
-            w.to_string(),
-            "stale mitmproxy PID 555 on port 32821 (runner stopped)"
-        );
-
-        let w = Warning::OrphanNbdDevice {
-            device_index: 3,
-            pid: 12345,
-        };
-        assert_eq!(
-            w.to_string(),
-            "orphan NBD device /dev/nbd3 (owner PID 12345 no longer exists)"
-        );
-
-        let w = Warning::NbdScanFailed;
-        assert_eq!(w.to_string(), "NBD orphan scan failed (task panicked)");
-    }
-
-    #[test]
-    fn is_inactive_mode_classification() {
-        assert!(is_inactive_mode("stopped"));
-        assert!(is_inactive_mode("draining"));
-        assert!(!is_inactive_mode("running"));
-        assert!(!is_inactive_mode("starting"));
-        assert!(!is_inactive_mode(""));
-    }
-
-    #[test]
-    fn idle_vm_diagnostic_line_includes_session_id() {
-        let raw_session_id = "sess-sensitive-doctor-17975";
-        let line = format_idle_vm_diagnostic_line(&IdleVm {
-            session_id: raw_session_id.into(),
-            sandbox_id: "sandbox-123".into(),
-        });
-
-        assert!(line.contains(raw_session_id));
-        assert!(line.contains("sandbox-123"));
-    }
-
-    struct DoctorReportFixture {
-        _dir: tempfile::TempDir,
-        config_path: PathBuf,
-        base_dir: PathBuf,
-    }
-
-    fn doctor_report_fixture(
-        mode: &str,
-        proxy_port: Option<u16>,
-        dns_port: Option<u16>,
-    ) -> DoctorReportFixture {
-        let dir = tempfile::tempdir().unwrap();
-        let base_dir = dir.path().join("runner");
-        std::fs::create_dir_all(&base_dir).unwrap();
-
-        let config_path = dir.path().join("runner.yaml");
-        let config = serde_json::json!({
-            "name": "test-runner",
-            "group": "vm0/test",
-            "base_dir": base_dir.display().to_string(),
-            "ca_dir": dir.path().join("ca").display().to_string(),
-            "firecracker": {
-                "binary": dir.path().join("firecracker").display().to_string(),
-                "kernel": dir.path().join("vmlinux").display().to_string(),
-            },
-            "profiles": {
-                "vm0/default": {
-                    "rootfs_hash": "rootfs",
-                    "snapshot_hash": "snapshot",
-                    "vcpu": 1,
-                    "memory_mb": 512,
-                    "rootfs_disk_mb": 512,
-                    "workspace_disk_mb": 1024,
-                },
-            },
-        });
-        std::fs::write(&config_path, serde_yaml_ng::to_string(&config).unwrap()).unwrap();
-
-        let status = serde_json::json!({
-            "mode": mode,
-            "started_at": "2026-01-01T00:00:00.000Z",
-            "active_runs": [],
-            "proxy_port": proxy_port,
-            "dns_port": dns_port,
-        });
-        std::fs::write(base_dir.join("status.json"), status.to_string()).unwrap();
-
-        DoctorReportFixture {
-            _dir: dir,
-            config_path,
-            base_dir,
-        }
-    }
-
-    fn live_runner_instance(
-        pid: u32,
-        config_path: PathBuf,
-        base_dir: PathBuf,
-    ) -> LiveRunnerInstance {
-        LiveRunnerInstance {
-            pid,
-            starttime: 0,
-            config_path,
-            base_dir,
-            runner_name: "test-runner".into(),
-            runner_group: "vm0/test".into(),
-            subcommand: "start".into(),
-            started_at: "2026-01-01T00:00:00.000Z".into(),
-        }
-    }
-
-    async fn build_test_runner_report(
-        mode: &str,
-        proxy_port: Option<u16>,
-        dns_port: Option<u16>,
-        mitm_procs: Vec<process::MitmproxyProcessInfo>,
-        dns_procs: Vec<process::DnsmasqProcessInfo>,
-    ) -> RunnerReport {
-        let fixture = doctor_report_fixture(mode, proxy_port, dns_port);
-        let runner = live_runner_instance(
-            std::process::id(),
-            fixture.config_path.clone(),
-            fixture.base_dir.clone(),
-        );
-        let report = build_runner_report(&runner, &[], &mitm_procs, &dns_procs, &[]).await;
-        assert!(report.status.is_some(), "test status should load");
-        report
-    }
-
-    fn mitm_proc(pid: u32, port: u16) -> process::MitmproxyProcessInfo {
-        process::MitmproxyProcessInfo {
-            pid,
-            ppid: None,
-            port,
-        }
-    }
-
-    fn dns_proc(pid: u32, port: u16) -> process::DnsmasqProcessInfo {
-        process::DnsmasqProcessInfo { pid, port }
-    }
-
-    fn empty_discovered() -> process::DiscoveredProcesses {
-        process::DiscoveredProcesses {
-            firecrackers: vec![],
-            mitmdumps: vec![],
-            dnsmasqs: vec![],
-        }
-    }
-
-    fn has_proxy_warning(report: &RunnerReport) -> bool {
-        report.warnings.iter().any(|w| {
-            matches!(
-                w,
-                Warning::NoMitmproxy { .. } | Warning::StaleMitmproxy { .. }
-            )
-        })
-    }
-
-    fn has_dns_warning(report: &RunnerReport) -> bool {
-        report
-            .warnings
-            .iter()
-            .any(|w| matches!(w, Warning::NoDnsmasq { .. }))
-    }
-
-    #[tokio::test]
-    async fn recheck_clears_runner_warnings_when_registry_entry_disappears() {
-        let dir = tempfile::tempdir().unwrap();
-        let home = HomePaths::with_root(dir.path().join("vm0-runner"));
-        let base_dir = dir.path().join("runner");
-        let runner = live_runner_instance(
-            std::process::id(),
-            dir.path().join("runner.yaml"),
-            base_dir.clone(),
-        );
-        let mut reports = vec![RunnerReport {
-            live_runner: runner.clone(),
-            service_type: ServiceType::Bare,
-            status: None,
-            api_ok: None,
-            proxy_pid: None,
-            dns_pid: None,
-            jobs: vec![],
-            warnings: vec![Warning::NoMitmproxy {
-                port: 32821,
-                base_dir,
-            }],
-        }];
-
-        recheck_per_runner_warnings(&home, &empty_discovered(), &mut reports)
-            .await
-            .unwrap();
-
-        assert!(reports[0].warnings.is_empty());
-    }
-
-    #[tokio::test]
-    async fn recheck_keeps_runner_warnings_while_registry_entry_is_current() {
-        let dir = tempfile::tempdir().unwrap();
-        let home = HomePaths::with_root(dir.path().join("vm0-runner"));
-        let base_dir = dir.path().join("runner");
-        std::fs::create_dir_all(&base_dir).unwrap();
-        std::fs::write(
-            base_dir.join("status.json"),
-            r#"{
-                "mode": "running",
-                "active_runs": [],
-                "started_at": "2026-01-01T00:00:00.000Z",
-                "proxy_port": 32821
-            }"#,
-        )
-        .unwrap();
-        let _handle = crate::live_runner_instances::publish(
-            &home,
-            crate::live_runner_instances::LiveRunnerInstanceMetadata {
-                config_path: dir.path().join("runner.yaml"),
-                base_dir: base_dir.clone(),
-                runner_name: "test-runner".into(),
-                runner_group: "vm0/test".into(),
-                subcommand: "start".into(),
-            },
-        )
-        .await
-        .unwrap();
-        let runner = crate::live_runner_instances::try_list(&home)
-            .await
-            .unwrap()
-            .into_iter()
-            .next()
-            .unwrap();
-        let mut reports = vec![RunnerReport {
-            live_runner: runner.clone(),
-            service_type: ServiceType::Bare,
-            status: None,
-            api_ok: None,
-            proxy_pid: None,
-            dns_pid: None,
-            jobs: vec![],
-            warnings: vec![Warning::NoMitmproxy {
-                port: 32821,
-                base_dir,
-            }],
-        }];
-
-        recheck_per_runner_warnings(&home, &empty_discovered(), &mut reports)
-            .await
-            .unwrap();
-
-        assert_eq!(reports[0].warnings.len(), 1);
-    }
-
-    #[tokio::test]
-    async fn recheck_fails_when_live_registry_entry_becomes_invalid() {
-        let dir = tempfile::tempdir().unwrap();
-        let home = HomePaths::with_root(dir.path().join("vm0-runner"));
-        let base_dir = dir.path().join("runner");
-        let _handle = crate::live_runner_instances::publish(
-            &home,
-            crate::live_runner_instances::LiveRunnerInstanceMetadata {
-                config_path: dir.path().join("runner.yaml"),
-                base_dir: base_dir.clone(),
-                runner_name: "test-runner".into(),
-                runner_group: "vm0/test".into(),
-                subcommand: "start".into(),
-            },
-        )
-        .await
-        .unwrap();
-        let runner = crate::live_runner_instances::try_list(&home)
-            .await
-            .unwrap()
-            .into_iter()
-            .next()
-            .unwrap();
-        crate::state_file::write_private_atomic(
-            &home.live_runner_instance_record_path(runner.pid, runner.starttime),
-            b"{",
-        )
-        .await
-        .unwrap();
-        let mut reports = vec![RunnerReport {
-            live_runner: runner.clone(),
-            service_type: ServiceType::Bare,
-            status: None,
-            api_ok: None,
-            proxy_pid: None,
-            dns_pid: None,
-            jobs: vec![],
-            warnings: vec![Warning::NoMitmproxy {
-                port: 32821,
-                base_dir,
-            }],
-        }];
-
-        let error =
-            match recheck_per_runner_warnings(&home, &empty_discovered(), &mut reports).await {
-                Ok(_) => panic!("expected invalid live registry entry to fail"),
-                Err(error) => error,
-            };
-
-        assert!(
-            error.to_string().contains("live process identity"),
-            "{error}"
-        );
-    }
-
-    #[tokio::test]
-    async fn report_uses_registry_identity_when_config_is_missing() {
-        let dir = tempfile::tempdir().unwrap();
-        let base_dir = dir.path().join("runner");
-        std::fs::create_dir_all(&base_dir).unwrap();
-        std::fs::write(
-            base_dir.join("status.json"),
-            r#"{
-                "mode": "running",
-                "max_concurrent": 4,
-                "active_runs": [],
-                "started_at": "2026-01-01T00:00:00.000Z",
-                "updated_at": "2026-01-01T00:00:00.000Z"
-            }"#,
-        )
-        .unwrap();
-        let runner = live_runner_instance(
-            std::process::id(),
-            dir.path().join("missing-runner.yaml"),
-            base_dir.clone(),
-        );
-
-        let report = build_runner_report(&runner, &[], &[], &[], &[]).await;
-
-        assert_eq!(report.live_runner.runner_name, "test-runner");
-        assert_eq!(report.live_runner.base_dir, base_dir);
-        assert_eq!(
-            report.live_runner.config_path,
-            dir.path().join("missing-runner.yaml")
-        );
-        assert_eq!(report.live_runner.pid, std::process::id());
-        assert_eq!(report.live_runner.subcommand, "start");
-        assert!(report.status.is_some());
-        assert_eq!(report.api_ok, None);
-    }
-
-    #[tokio::test]
-    async fn report_uses_registry_subcommand() {
-        let fixture = doctor_report_fixture("running", None, None);
-        let mut runner = live_runner_instance(
-            std::process::id(),
-            fixture.config_path.clone(),
-            fixture.base_dir.clone(),
-        );
-        runner.subcommand = "benchmark".into();
-
-        let report = build_runner_report(&runner, &[], &[], &[], &[]).await;
-
-        assert_eq!(report.live_runner.subcommand, "benchmark");
-    }
-
-    #[tokio::test]
-    async fn build_runner_reports_requires_live_registry_entries() {
-        let server = MockServer::start_async().await;
-        let api = server
-            .mock_async(|when, then| {
-                when.method("HEAD").path("/api");
-                then.status(200);
-            })
-            .await;
-        let fixture = doctor_report_fixture("running", None, None);
-        let config = serde_json::json!({
-            "name": "spoofed-runner",
-            "group": "vm0/test",
-            "base_dir": fixture.base_dir.display().to_string(),
-            "ca_dir": fixture._dir.path().join("ca").display().to_string(),
-            "firecracker": {
-                "binary": fixture._dir.path().join("firecracker").display().to_string(),
-                "kernel": fixture._dir.path().join("vmlinux").display().to_string(),
-            },
-            "profiles": {
-                "vm0/default": {
-                    "rootfs_hash": "rootfs",
-                    "snapshot_hash": "snapshot",
-                    "vcpu": 1,
-                    "memory_mb": 512,
-                    "rootfs_disk_mb": 512,
-                    "workspace_disk_mb": 1024,
-                },
-            },
-            "server": {
-                "url": server.url("/api"),
-                "token": "spoofed-token",
-            },
-        });
-        std::fs::write(
-            &fixture.config_path,
-            serde_yaml_ng::to_string(&config).unwrap(),
-        )
-        .unwrap();
-        let discovered = empty_discovered();
-
-        let reports = build_runner_reports(&[], &discovered, &[]).await;
-
-        assert!(reports.is_empty());
-        api.assert_calls_async(0).await;
-    }
-
-    #[tokio::test]
-    async fn report_warns_for_running_without_proxy() {
-        let report = build_test_runner_report("running", Some(32821), None, vec![], vec![]).await;
-        assert!(
-            report
-                .warnings
-                .iter()
-                .any(|w| matches!(w, Warning::NoMitmproxy { port: 32821, .. }))
-        );
-    }
-
-    #[tokio::test]
-    async fn report_warns_stale_proxy_on_stopped_runner() {
-        let report = build_test_runner_report(
-            "stopped",
-            Some(32821),
-            None,
-            vec![mitm_proc(999, 32821)],
-            vec![],
-        )
-        .await;
-        assert!(report.warnings.iter().any(|w| matches!(
-            w,
-            Warning::StaleMitmproxy {
-                pid: 999,
-                port: 32821,
-            }
-        )));
-    }
-
-    #[tokio::test]
-    async fn report_no_warning_for_stopped_without_proxy() {
-        let report = build_test_runner_report("stopped", Some(32821), None, vec![], vec![]).await;
-        assert!(!has_proxy_warning(&report));
-    }
-
-    #[tokio::test]
-    async fn report_no_warning_for_draining_proxy() {
-        let without_proxy =
-            build_test_runner_report("draining", Some(32821), None, vec![], vec![]).await;
-        assert!(!has_proxy_warning(&without_proxy));
-
-        let with_proxy = build_test_runner_report(
-            "draining",
-            Some(32821),
-            None,
-            vec![mitm_proc(999, 32821)],
-            vec![],
-        )
-        .await;
-        assert!(!has_proxy_warning(&with_proxy));
-    }
-
-    #[tokio::test]
-    async fn report_no_proxy_or_dns_warning_while_starting() {
-        let report =
-            build_test_runner_report("starting", Some(32821), Some(5353), vec![], vec![]).await;
-
-        assert!(!has_proxy_warning(&report));
-        assert!(!has_dns_warning(&report));
-    }
-
-    #[tokio::test]
-    async fn report_no_warning_for_running_with_proxy() {
-        let report = build_test_runner_report(
-            "running",
-            Some(32821),
-            None,
-            vec![mitm_proc(999, 32821)],
-            vec![],
-        )
-        .await;
-        assert!(!has_proxy_warning(&report));
-    }
-
-    #[tokio::test]
-    async fn report_warns_for_running_when_only_unrelated_proxy_exists() {
-        let report = build_test_runner_report(
-            "running",
-            Some(32821),
-            None,
-            vec![mitm_proc(999, 32822)],
-            vec![],
-        )
-        .await;
-        assert!(
-            report
-                .warnings
-                .iter()
-                .any(|w| matches!(w, Warning::NoMitmproxy { port: 32821, .. }))
-        );
-    }
-
-    #[tokio::test]
-    async fn report_no_stale_proxy_warning_for_stopped_with_unrelated_proxy() {
-        let report = build_test_runner_report(
-            "stopped",
-            Some(32821),
-            None,
-            vec![mitm_proc(999, 32822)],
-            vec![],
-        )
-        .await;
-        assert!(!has_proxy_warning(&report));
-    }
-
-    #[tokio::test]
-    async fn report_warns_for_running_without_dnsmasq() {
-        let report = build_test_runner_report("running", None, Some(5353), vec![], vec![]).await;
-        assert!(
-            report
-                .warnings
-                .iter()
-                .any(|w| matches!(w, Warning::NoDnsmasq { port: 5353, .. }))
-        );
-    }
-
-    #[tokio::test]
-    async fn report_no_warning_for_running_with_dnsmasq() {
-        let report = build_test_runner_report(
-            "running",
-            None,
-            Some(5353),
-            vec![],
-            vec![dns_proc(888, 5353)],
-        )
-        .await;
-        assert!(!has_dns_warning(&report));
-    }
-
-    #[tokio::test]
-    async fn report_warns_for_running_when_only_unrelated_dnsmasq_exists() {
-        let report = build_test_runner_report(
-            "running",
-            None,
-            Some(5353),
-            vec![],
-            vec![dns_proc(888, 5354)],
-        )
-        .await;
-        assert!(
-            report
-                .warnings
-                .iter()
-                .any(|w| matches!(w, Warning::NoDnsmasq { port: 5353, .. }))
-        );
-    }
-
-    #[test]
-    fn is_test_tld_matches_dot_test() {
-        assert!(is_test_tld("https://not-a-real-server.test/api"));
-        assert!(is_test_tld("https://sub.domain.test"));
-        assert!(is_test_tld("https://test"));
-        assert!(is_test_tld("https://server.test:8080/api"));
-    }
-
-    #[test]
-    fn is_test_tld_rejects_substring_match() {
-        assert!(!is_test_tld("https://attestation.service.internal/api"));
-        assert!(!is_test_tld("https://my.testing.company.com/api"));
-        assert!(!is_test_tld("https://contest.example.com"));
-    }
-
-    #[test]
-    fn is_test_tld_handles_edge_cases() {
-        assert!(!is_test_tld("not-a-url"));
-        assert!(!is_test_tld("https://example.com/.test"));
-        assert!(!is_test_tld("https://example.com?q=.test"));
-    }
-
-    #[tokio::test]
-    async fn is_lock_free_returns_true_when_file_not_found() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("no-such-file.lock");
-        assert!(super::is_lock_free(path.to_str().unwrap()).await);
-    }
-
-    #[tokio::test]
-    async fn is_lock_free_returns_true_when_lock_not_held() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("free.lock");
-        std::fs::File::create(&path).unwrap();
-        assert!(super::is_lock_free(path.to_str().unwrap()).await);
-    }
-
-    #[tokio::test]
-    async fn is_lock_free_returns_false_when_lock_held() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("held.lock");
-        let file = std::fs::File::create(&path).unwrap();
-        // Hold an exclusive lock for the duration of the test.
-        let _lock = nix::fcntl::Flock::lock(file, nix::fcntl::FlockArg::LockExclusiveNonblock)
-            .expect("failed to acquire test lock");
-        assert!(!super::is_lock_free(path.to_str().unwrap()).await);
-    }
-}
+mod tests;
