+++ crates/runner/src/cmd/doctor/tests/mod.rs
@@ -1,0 +1,7 @@
+mod correlation;
+mod formatting;
+mod host;
+mod recheck;
+mod report;
+mod status;
+mod support;
